// redsocial-frontend/src/interfaces/publicationInterface.ts
// Generated by https://quicktype.io

export interface IPublication {
    status?:             string;
    message?:            string;
    page:               number;
    pages:              number;
    total:              number;
    publicationsByUser: PublicationsByUser[];
}

export interface PublicationsByUser {
    _id:        string;
    text:       string;
    created_at: string;
    user:       User;
    __v:        number;
}

export interface User {
    _id:       string;
    name:      string;
    surname:   string;
    nick:      string;
    image:     string;
    create_at: string;
    bio:       string;
}